Refactor file layout
**Why**: All our tests are feature tests with js: true so configure
that at the top-level. Break the helpers up into smaller files.
